Add option to get label value from an HTTP header
This allows to use the proxy in settings where query parameters
cannot be used; this could be behind specific reverse proxies, or
when HTTP POST is required (see #53).

This work is heavily based on the work done by
Andreas Plesner <apj@mutt.dk> in [1]. I (Jonas Schäfer) ported
this to the current head of master.

   [1]: https://github.com/prometheus-community/prom-label-proxy/pull/54

Signed-off-by: Jonas Schäfer <j.wielicki@sotecware.net>