go,vendor: bump Prometheus dependency
This commit bumps the dependency on Prometheus to the commit from the
v2.28.0 tag
(https://github.com/prometheus/prometheus/releases/tag/v2.28.0). The
reason for bumping this is to take advantage of the updates to the
PromQL parser. Without this update, the prom-label-proxy cannot enforce
labels on queries that use any new features, e.g. queries that include:
* the last_over_time, sgn, clamp functions, added in
  https://github.com/prometheus/prometheus/pull/8457;
* the @ operator, added in
  https://github.com/prometheus/prometheus/pull/8121; or
* the group aggregation operator, added in
  https://github.com/prometheus/prometheus/pull/7480.

Currently, queries including these functions produce errors like:
`parse error: unknown function with name "group"`

In order to keep tests working, we need to pull in the latest commits
from Alertmanager. This, however, does not satisfy `go mod` since
v0.22.2 of Alertmanager, on which Prometheus depends, is not merged into
main. As a result, we need a replace directive in the go.mod file for
now.

Signed-off-by: Lucas Servén Marín <lserven@gmail.com>