return error if existing label matcher in query would change (#67)
* Add option to return error if existing label matcher in query would change

Before this commit, if the inject proxy is configured to replace label
foo and a query is encountered that has one or more label matchers for
the foo label, these matchers are silently replaced. E.g. a query like
`up{foo="default"}` would silently be altered to
`up{foo="replacement"}` and the query result returned.
With this change we can return 400 instead. The --error-on-replace option
enables this behavior.

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>

* Don't capitalize error message

Co-authored-by: Lucas Servén Marín <lserven@gmail.com>
Signed-off-by: Jan Fajerski <jfajersk@redhat.com>

* injectproxy: report on labels when replacement causes error

For better error reporting, this commit changes the error returned when
a label is replaced and --error-on-replace was passed.

Signed-off-by: Jan Fajerski <jfajersk@redhat.com>

Co-authored-by: Lucas Servén Marín <lserven@gmail.com>