Do not overwrite existing filter when multi-values (Prometheus)
When prom-label-proxy is configured with multiple label values, it
should preserve the existing matchers because users actually want to
return results for a subset of the metrics. In practice when configured
with `namespace=~"bar|foo"`, the query `up{namespace="foo"}` is
translated into `up{namespace="foo",namespace=~"bar|foo"}` instead of
`{namespace=~"bar|foo"}`.

This change is similar to 26b0a611 but for the Prometheus API endpoints.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>