feat: allow to return rules with matching active alerts
This change adds a `-rules-with-active-alerts` CLI argument to return
rules with active alerts matching the enforced label. The use case is
for an admin to write general-purpose alerting rules which individual
users can still retrieve from the API.

To avoid surprises ("why do I see this alert now?"), the feature needs to
be explicitly enabled.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>